Move RandomAccessFileInputStream/OutputStream to :files
