+++ platforms/software/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/resolveengine/store/DefaultBinaryStore.java
@@ -17,7 +17,7 @@ package org.gradle.api.internal.artifacts.ivyservice.resolveengine.store;
 
 import org.gradle.cache.internal.BinaryStore;
 import org.gradle.internal.concurrent.CompositeStoppable;
-import org.gradle.internal.io.RandomAccessFileInputStream;
+import org.gradle.internal.file.RandomAccessFileInputStream;
 import org.gradle.internal.serialize.Decoder;
 import org.gradle.internal.serialize.kryo.StringDeduplicatingKryoBackedDecoder;
 import org.gradle.internal.serialize.kryo.StringDeduplicatingKryoBackedEncoder;
